improved PROGMEM Glyph handling
Refactored the Glyph class to introduce a concept of Mutable. When inmutable,
the Glyph can simply wrap a PROGMEM block without incuring addition RAM. But
inmutable Glyphs are not required to be PROGMEM memory blocks either.

Update the space invaders example to us a PROGMEM block to initialize a inmutable
Glyph object.