removed flicker from scrolling animations
Removed the flickerthat occured during scrolling animations when the
animated image or glyph was not the same size as the LED matrix.
Also made the start/stop drawing calls on the matrix object stackable.
Each startDrawing() must be balanced with a stopDrawing() call to enable
screen updates for the changes.