improved update cycle timing
The last row of the matrix ended up being marginally brighter than the others.
The root cause was the scan row management that occured after the last row update.
On the last row, that management code took a longer code path to reset various
counter. This of course took longer and as a result the last row was on for a marginally
longer period of time than the other rows. This timing difference was perceptible in
the red color on the matrix with the red in the last row being a little bit brighter
than the other rows.

Solution was to move the scan row management code to outside of the code path
where the update timer is turned off. By doing so, the work done when the update
timer is off is now consistent for each row update, making the timing of each update
consistent and thus the brightness of each row consistent.